requirements.txt
Flask==2.2.3
Jinja2==3.1.2
requests==2.28.2
urllib3==1.26.15
Flask-OAuthlib==0.9.6
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.2.2
spotipy==2.23.0
matplotlib==3.6.3
pandas==1.5.3
requests==2.28.2
